+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * insert / delete in MySQL
 * relation tags {{ Variable }} to help generate compound sentences
 * allow for left recursion in generation grammar
+* local variables; let
 
 ## 1.11 modules - 13-09-2020